(30/10) Trocando axios por ky
