+++ package.json
@@ -12,6 +12,7 @@
     "@mui/x-date-pickers": "^5.0.0-alpha.4",
     "axios": "^0.27.2",
     "date-fns": "^2.28.0",
+    "ky": "^0.32.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-input-mask": "^2.0.4",
