+++ src/api.js
@@ -1,14 +1,9 @@
-import axios from 'axios'
+import ky from 'ky'
 
-const api = axios.create({
-    // Precisa terminar com /
-    baseURL: 'https://agoravai-fausto.onrender.com/',
-    timeout: 10000,
-    // Envia os cookies seguros de volta em todas as requisições
-    withCredentials: true
-    /*headers: {
-        'x-access-token': window.sessionStorage.getItem('app-data')
-    }*/
+const api = ky.extend({
+    prefixUrl: 'https://agoravai-fausto.onrender.com/',
+    credentials: true,
+    timeout: 10000
 })
 
 export default api
